Installs each service with a Helm chart (#2)
TL;DR
-----

Switches from templated manifests to Helm to install services

Details
-------

Uses the available Helm chart to deploy the Online Boutique, but
in a unique way. Each service is installed as it's own instance of
the Helm chart that installs just that service. This is possible 
since the chart allows toggling each service independently.

The single service/single chart paradigm allows for demonstrating
a workflow in which each service team is responsible for delivering
independently. Team build their apps and a chart, then a common
pipeline can pull the various releases together into an application
release. In this model, think of each instance of the chart as
an independent Helm chart.