Implements an application release pipeline (#3)
TL;DR
-----

Releases with Replicated when any service updates

Details
-------

Implements continuous delivered to an `edge` channel on the
Replicated vendor portal whenever a new image is released for any
of the Online Boutique microservices. The release pipeline
watches for new images being published, validates their
signatures, then creates a new Replicated release. This allows
end customers who are frequently polling for updates to receive
the latest updates as soon as they're released.

The release pipeline pins images in each Replicated release to
the hash of the the image. The release pipeline update the Helm
Chart object for the release to assure the right hash is used. To
provide traceability, each update has a PR associated with it
that is automatically approved if the PR head is has the expected
checksum. If additional commits have been added to the branch
since the PR was opened, the release fails assuming that was
something suspect.

All releases have semantic version tracked using the Concourse
semver resource.