Renames mis-named shipping service chart
TL;DR
-----

Corrects typo in shipping service chart manifest filename

Details
-------

Repairs pipeline error caused by typo in the name of the manifest file
for the Helm Chart manifest for the Shipping Service. The word `service`
was mis-spelled causing the `prepare-release` job to fail when there
was a change to the shipping service.