Deploys to notional SaaS cluster with Helm (#26)
TL;DR
-----

Demonstrates parallel deployment to SaaS cluster and vendor portal

Details
-------

Adds in a deployment directly to a customer that represents a
notional SaaS environment. This allows for demonstrating the parallel
flow of a new release to the Vendor portal and the running SaaS 
cluster.

The "SaaS" deploy is run using Helm with a post-renderer that 
customizes the Online Boutique Helm chart to use the latest 
image digests.