Corrects commit message and PR text (#34)
TL;DR
-----

Updates commit message and PR text to say what really happened

Details
-------

Repairs a defect where the original commit message and PR text
on the release update was always referencing the Ad Service. That
was an artifact of validating the pipeline with an implementation
that only changed the Ad Serviec and not going back to update the
hard coded texts.

I found the error while doing a walkthrough where I was updating
the frontend as part of a demo and noticing that I had a commit
message and PR referring to the ad service.

The new message and PR text refer to the commit/PR as and automated
service update add show a bulleted list of the services that 
have been updated.